// src/modules/property/enums/property-category.enum.ts
export enum EnumPropertyCategory {
  APARTMENT_SALE = 'APARTMENT_SALE',   // Kvartira SOTISH
  APARTMENT_RENT = 'APARTMENT_RENT',   // Kvartira IJARA
  COMMERCIAL_SALE = 'COMMERCIAL_SALE', // Noturar mulk SOTISH
  COMMERCIAL_RENT = 'COMMERCIAL_RENT', // Noturar mulk IJARA
  LAND_SALE = 'LAND_SALE',             // Yer SOTISH
  LAND_RENT = 'LAND_RENT',             // Yer IJARA
  GARAGE_SALE = 'GARAGE_SALE',         // Garaj SOTISH
  GARAGE_RENT = 'GARAGE_RENT',         // Garaj IJARA
  HOVLI_SALE = 'HOVLI_SALE',           // Hovli SOTISH
  HOVLI_RENT = 'HOVLI_RENT',           // Hovli IJARA
}